Fix legacy package 한쪽으로 몰아넣음
