+++ src/main/java/com/moment/the/admin/service/AdminServiceImpl.java
@@ -5,8 +5,8 @@ import com.moment.the.admin.dto.AdminDto;
 import com.moment.the.admin.dto.SignInDto;
 import com.moment.the.admin.repository.AdminRepository;
 import com.moment.the.config.security.jwt.JwtUtil;
-import com.moment.the.exception.legacyException.UserAlreadyExistsException;
-import com.moment.the.exception.legacyException.UserNotFoundException;
+import com.moment.the.exception.legacy.legacyException.UserAlreadyExistsException;
+import com.moment.the.exception.legacy.legacyException.UserNotFoundException;
 import com.moment.the.util.RedisUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.context.SecurityContextHolder;
