+++ src/main/java/com/moment/the/exception/legacy/legacyException/AccessNotFoundException.java
@@ -1,4 +1,4 @@
-package com.moment.the.exception.legacyException;
+package com.moment.the.exception.legacy.legacyException;
 
 public class AccessNotFoundException extends RuntimeException{
     public AccessNotFoundException(String msg, Throwable t){
