+++ src/main/java/com/moment/the/exception/legacy/legacyException/NoGoodException.java
@@ -1,4 +1,4 @@
-package com.moment.the.exception.legacyException;
+package com.moment.the.exception.legacy.legacyException;
 
 public class NoGoodException extends RuntimeException{
     public NoGoodException(String msg, Throwable t){
