+++ src/main/java/com/moment/the/exception/legacy/legacyException/NoImprovementException.java
@@ -1,4 +1,4 @@
-package com.moment.the.exception.legacyException;
+package com.moment.the.exception.legacy.legacyException;
 
 public class NoImprovementException extends RuntimeException{
     public NoImprovementException(String msg, Throwable t){
