+++ src/main/java/com/moment/the/exception/legacy/legacyHandler/AccessTokenExpiredHandler.java
@@ -1,15 +1,12 @@
-package com.moment.the.exception.legacyHandler;
-
-import lombok.extern.slf4j.Slf4j;
+package com.moment.the.exception.legacy.legacyHandler;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@Slf4j
-public class NoPostHandler {
+public class AccessTokenExpiredHandler {
     public void handle(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
-        response.sendRedirect("/exception/noPost");
+        response.sendRedirect("/exception/access-token-expired");
     }
 }
