+++ src/main/java/com/moment/the/exception/legacy/legacyHandler/AnswerAlreadyExistsHandler.java
@@ -1,12 +1,13 @@
-package com.moment.the.exception.legacyHandler;
+package com.moment.the.exception.legacy.legacyHandler;
+
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class InvalidTokenHandler {
+public class AnswerAlreadyExistsHandler {
     public void handle(HttpServletRequest req, HttpServletResponse res) throws IOException, ServletException {
-        res.sendRedirect("/exception/invalid-token");
+        res.sendRedirect("/exception/answer-already-exists");
     }
 }
