+++ src/main/java/com/moment/the/exception/legacy/legacyHandler/InvalidTokenHandler.java
@@ -1,12 +1,12 @@
-package com.moment.the.exception.legacyHandler;
+package com.moment.the.exception.legacy.legacyHandler;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class AccessNotFoundHandler {
+public class InvalidTokenHandler {
     public void handle(HttpServletRequest req, HttpServletResponse res) throws IOException, ServletException {
-        res.sendRedirect("/exception/access-not-found");
+        res.sendRedirect("/exception/invalid-token");
     }
 }
