+++ src/main/java/com/moment/the/improvement/service/ImprovementService.java
@@ -3,8 +3,8 @@ package com.moment.the.improvement.service;
 import com.moment.the.admin.AdminDomain;
 import com.moment.the.admin.repository.AdminRepository;
 import com.moment.the.admin.service.AdminServiceImpl;
-import com.moment.the.exception.legacyException.AccessNotFoundException;
-import com.moment.the.exception.legacyException.UserNotFoundException;
+import com.moment.the.exception.legacy.legacyException.AccessNotFoundException;
+import com.moment.the.exception.legacy.legacyException.UserNotFoundException;
 import com.moment.the.improvement.ImprovementDomain;
 import com.moment.the.improvement.dto.ImprovementDto;
 import com.moment.the.improvement.dto.ImprovementViewAllDto;
