+++ src/main/java/com/moment/the/uncomfortable/service/UncomfortableService.java
@@ -1,6 +1,6 @@
 package com.moment.the.uncomfortable.service;
 
-import com.moment.the.exception.legacyException.NoPostException;
+import com.moment.the.exception.legacy.legacyException.NoPostException;
 import com.moment.the.uncomfortable.UncomfortableDomain;
 import com.moment.the.uncomfortable.dto.UncomfortableResponseDto;
 import com.moment.the.uncomfortable.dto.UncomfortableSetDto;
