+++ src/test/java/com/moment/the/service/AnswerServiceTest.java
@@ -1,7 +1,7 @@
 package com.moment.the.service;
 
-import com.moment.the.exception.legacyException.AccessNotFoundException;
-import com.moment.the.exception.legacyException.AnswerAlreadyExistsException;
+import com.moment.the.exception.legacy.legacyException.AccessNotFoundException;
+import com.moment.the.exception.legacy.legacyException.AnswerAlreadyExistsException;
 import com.moment.the.answer.*;
 import com.moment.the.answer.dto.AnswerDto;
 import com.moment.the.answer.dto.AnswerResDto;
